Abstract out ForeignModConverter.
No intentional functional changes.
